+++ apis/supernova/package.json
@@ -5,7 +5,6 @@
   "main": "src/index.js",
   "devDependencies": {
     "extend": "3.0.2",
-    "node-event-emitter": "0.0.1",
-    "@nebula.js/nucleus": "^5.12.4"
+    "node-event-emitter": "0.0.1"
   }
 }
