+++ apis/supernova/src/handler/__test__/hypercube-handler.test.js
@@ -1,4 +1,4 @@
-import HyperCubeHandler from '../handlers/hypercube-handler';
+import HyperCubeHandler from '../hypercube-handler';
 
 describe('HyperCube Handlers', () => {
   let handler;
